+++ client/src/components/LayOut.jsx
@@ -1,11 +1,14 @@
 import { Outlet } from "react-router-dom";
-import Navbar from "./NavBar";
+import Navbar from "./Navbar/Navbar";
+import Footer from "./Footer/Footer";
+
 
 export const LayOut = () => {
   return (
     <div>
       <Navbar />
       <Outlet />
+      <Footer />
     </div>
   );
 };
