+++ client/src/pages/adminPage.jsx
@@ -1,21 +1,21 @@
-import React, {useEffect, useRef} from "react";
-import {useState} from "react";
-import Axios from "axios";
-import {useSelector, useDispatch} from "react-redux";
-import {Link} from "react-routerdom";
+// import React, {useEffect, useRef} from "react";
+// import {useState} from "react";
+// import Axios from "axios";
+// import {useSelector, useDispatch} from "react-redux";
+// import {Link} from "react-routerdom";
 
-import Dropdown from "react-bootstrap/Dropdown";
-import DropdownButton from "react-bootstrap/DropdownButton";
-import Button from "react-bootstrap/Button";
-import Table from "react-bootstrap/Table";
-import Navbar from "react-bootstrap/Navbar";
-import Container from "react-bootstrap/Container";
-import NavDropdown from "react-bootstrap/NavDropdown";
-import OverlayTrigger from "react-bootstrap/OverlayTrigger";
-import Popover from "react-bootstrap/Popover";
-import Form from "react-bootstrap/Form";
+// import Dropdown from "react-bootstrap/Dropdown";
+// import DropdownButton from "react-bootstrap/DropdownButton";
+// import Button from "react-bootstrap/Button";
+// import Table from "react-bootstrap/Table";
+// import Navbar from "react-bootstrap/Navbar";
+// import Container from "react-bootstrap/Container";
+// import NavDropdown from "react-bootstrap/NavDropdown";
+// import OverlayTrigger from "react-bootstrap/OverlayTrigger";
+// import Popover from "react-bootstrap/Popover";
+// import Form from "react-bootstrap/Form";
 
-import {} from "../" 
-export const adminPage = () => {
+// import {} from "../" 
+// export const adminPage = () => {
 
-}
+// }
